updated and optimised several methods
versão já a funcionar, todas as regras cumpridas.
falta limpar código

+++ projectoFinalPP_bank/src/models/Account.java
@@ -4,10 +4,10 @@ public interface Account {
 	
 	int getAccountId(); // Devolve o nº de conta
 	
-	double getAccountBalance(); // Devolve o valor em saldo na conta
+	int getAccountBalance(); // Devolve o valor em saldo na conta
 	
-	void accountTransferIn(double transferAmount); // Actualiza o saldo da conta de destino(transferência)
+	void accountTransferIn(int transferAmount); // Actualiza o saldo da conta de destino(transferência)
 	
-	void accountTransferOut(double transferAmount); // Actualiza o saldo da conta de origem(transferência)
+	void accountTransferOut(int transferAmount); // Actualiza o saldo da conta de origem(transferência)
 	
 }
